requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-nmcli==0.1.1
speedtest-cli==2.1.1
Werkzeug==0.15.5
